+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [unreleased]
 
 ### Changed
 
 - Unify crate naming scheme
 - Reduce boilerplate of introducing new logs/hdf5 files with some simple macros
+- Move supported format definitions to separate crate
 
 ## [1.14.2]